refactor: ImageSize 객체 생성하여 높이 너비 값 함께 관리하도록 수정
